services/meeting_service/meetings/tasks.py
from celery import shared_task
from django.core.files.storage import default_storage
from .models import MeetingRecording, Meeting
import logging

logger = logging.getLogger(__name__)

@shared_task
def process_meeting_recording(recording_id):
    """
    Process meeting recording after upload
    """
    try:
        recording = MeetingRecording.objects.get(id=recording_id)
        recording.status = 'processing'
        recording.save()

        # Implement video processing logic here
        # This would include:
        # - Video compilation
        # - Audio processing
        # - Thumbnail generation
        # - Format conversion

        recording.status = 'ready'
        recording.processing_completed_at = timezone.now()
        recording.save()

        logger.info(f"Recording {recording_id} processed successfully")

    except MeetingRecording.DoesNotExist:
        logger.error(f"Recording {recording_id} not found")
    except Exception as e:
        recording.status = 'failed'
        recording.save()
        logger.error(f"Failed to process recording {recording_id}: {str(e)}")

@shared_task
def send_meeting_reminder(meeting_id):
    """
    Send meeting reminder emails
    """
    try:
        meeting = Meeting.objects.get(id=meeting_id)
        participants = meeting.participants.all()

        for participant in participants:
            # Send reminder email
            # Implement email sending logic
            pass

    except Meeting.DoesNotExist:
        logger.error(f"Meeting {meeting_id} not found")

@shared_task
def cleanup_old_recordings():
    """
    Clean up old recordings based on retention policy
    """
    # Implement cleanup logic
    pass