services/meeting_service/requirements.txt
# Django Core
Django==5.2.1
djangorestframework==3.15.2
django-cors-headers==4.3.1
channels==4.1.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.2.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0

# Configuration
python-decouple==3.8

# Production
whitenoise==6.6.0
gunicorn==21.2.0

# Utilities
Pillow==10.3.0
python-dateutil==2.9.0
pytz==2024.1

